// src/app/models/Offer.ts
export class Offer {
    id: number;
    pairName: string;
    option1: string;
    optionX: string;
    option2: string;
    optionX1: string;
    optionX2: string;
    special: boolean;
    option1Selected: boolean;
    optionXSelected: boolean;
    option2Selected: boolean;
    optionX1Selected: boolean;
    optionX2Selected: boolean;
}